+++ priv/repo/migrations/20220706214259_create_items.exs
@@ -5,8 +5,8 @@ defmodule Anya.Repo.Migrations.CreateItems do
     create table(:items) do
       add :name, :string
       add :quantity, :integer
-      add :price, :integer
-      add :total, :integer
+      add :price, :float
+      add :total, :float
 
       add :invoice_id, references(:invoices, on_delete: :delete_all)
